refactor(physics): split rigid-body data classes into separate modules
Move RaycastResult, SingleContactResult, ContactPoint and ContactResult out of system.js into their own one-class modules, matching the engine's one-class-per-file convention. Behavior-neutral; the public API is unchanged.
